Move to `conversational_ai/` and split default_audio_interface
